refactor: break up CSharpTestController into focused modules (SRP)
Extract domain logic from the monolithic controller into dedicated
service modules so the controller only orchestrates.

- Extract result-matching logic into execution/resultMatcher.ts
- Rewrite execution/testRunner.ts to work with TestTreeNode
- Rewrite debug/debugLauncher.ts to work with TestTreeNode
- Extract status bar management into ui/statusBarManager.ts
- Slim testController.ts from 535 to ~170 lines of pure orchestration

No behavior change — all 79 existing tests pass.

Closes #12

Made-with: Cursor
