+++ src/debug/debugLauncher.ts
@@ -1,32 +1,26 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import { ChildProcess } from 'child_process';
-import { spawnDotnet, getExtraArgs } from '../utils/dotnetCli';
+import { TestTreeNode } from '../testTreeProvider';
+import { getExtraArgs } from '../utils/dotnetCli';
 import { log, logError, showOutput } from '../utils/outputChannel';
-import { buildFilter, getProjectPath } from '../execution/filterBuilder';
+import { buildFilterForNode } from '../execution/testRunner';
 
 const PID_REGEX = /Process Id:\s*(\d+)/;
 
-export async function debugTests(
-    testRun: vscode.TestRun,
-    items: readonly vscode.TestItem[],
+export async function launchDebugSession(
+    node: TestTreeNode,
     token: vscode.CancellationToken,
 ): Promise<void> {
-    if (items.length === 0) {
+    if (!node.projectPath) {
+        logError('No project path for node');
         return;
     }
 
-    const firstItem = items[0];
-    const projectPath = findProjectPath(firstItem);
-    if (!projectPath) {
-        logError('Cannot debug: no project path found for test item');
-        return;
-    }
+    const projectDir = path.dirname(node.projectPath);
+    const args = ['test', node.projectPath, '--no-restore'];
 
-    const projectDir = path.dirname(projectPath);
-    const { filter } = buildFilter(items);
-
-    const args = ['test', projectPath, '--no-restore'];
+    const filter = buildFilterForNode(node);
     if (filter) {
         args.push('--filter', filter);
     }
@@ -36,18 +30,14 @@ export async function debugTests(
         args.push(...extraArgs);
     }
 
-    log('Starting test host in debug mode (VSTEST_HOST_DEBUG=1)...');
+    log('Starting test host with VSTEST_HOST_DEBUG=1...');
     showOutput();
 
+    const { spawnDotnet } = await import('../utils/dotnetCli');
     const proc = spawnDotnet(args, projectDir, { VSTEST_HOST_DEBUG: '1' });
 
     try {
         const pid = await waitForPid(proc, token);
-
-        if (token.isCancellationRequested) {
-            proc.kill();
-            return;
-        }
 
         log(`Test host PID: ${pid}. Attaching debugger...`);
 
@@ -58,18 +48,15 @@ export async function debugTests(
             processId: pid.toString(),
         };
 
-        const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
-        const started = await vscode.debug.startDebugging(workspaceFolder, debugConfig);
+        const folder = vscode.workspace.workspaceFolders?.[0];
+        const started = await vscode.debug.startDebugging(folder, debugConfig);
 
         if (!started) {
-            logError(
-                'Failed to attach debugger. Make sure the C# extension (or OmniSharp) is installed.',
-            );
+            logError('Failed to attach debugger');
             proc.kill();
             return;
         }
 
-        // Wait for the process to exit
         await new Promise<void>((resolve) => {
             proc.on('close', () => resolve());
             token.onCancellationRequested(() => {
@@ -81,8 +68,8 @@ export async function debugTests(
         log('Debug session completed.');
     } catch (err) {
         proc.kill();
-        if (err instanceof Error && err.message !== 'Cancelled') {
-            logError('Debug session failed', err);
+        if (!(err instanceof Error && err.message === 'Cancelled')) {
+            logError('Debug failed', err);
         }
     }
 }
@@ -95,9 +82,8 @@ function waitForPid(proc: ChildProcess, token: vscode.CancellationToken): Promis
             buffer += data.toString();
             const match = buffer.match(PID_REGEX);
             if (match) {
-                const pid = parseInt(match[1], 10);
                 proc.stdout?.removeListener('data', onData);
-                resolve(pid);
+                resolve(parseInt(match[1], 10));
             }
         };
 
@@ -107,31 +93,14 @@ function waitForPid(proc: ChildProcess, token: vscode.CancellationToken): Promis
         });
 
         proc.on('close', (code) => {
-            reject(new Error(`Test host exited (code ${code}) before PID was detected`));
+            reject(new Error(`Test host exited (code ${code}) before PID detected`));
         });
 
-        const onCancel = token.onCancellationRequested(() => {
-            onCancel.dispose();
+        token.onCancellationRequested(() => {
+            proc.kill();
             reject(new Error('Cancelled'));
         });
 
-        // Timeout after 60 seconds
-        setTimeout(() => {
-            reject(
-                new Error('Timed out waiting for test host PID. Is VSTEST_HOST_DEBUG supported?'),
-            );
-        }, 60_000);
+        setTimeout(() => reject(new Error('Timeout waiting for test host PID')), 60_000);
     });
-}
-
-function findProjectPath(item: vscode.TestItem): string | undefined {
-    let current: vscode.TestItem | undefined = item;
-    while (current) {
-        const pp = getProjectPath(current);
-        if (pp) {
-            return pp;
-        }
-        current = current.parent;
-    }
-    return undefined;
 }
